+++ changelog.md
@@ -1,6 +1,8 @@
 ### 09.04.2023, Sun
 
 - 01:30 - Frontend: created bookmark filter by tags.
+- 19:30 - `bookmarks.vue` significantly refactored.
+- 19:50 - Added bookmark search by it's title.
 
 ### 08.04.2023, Sat